pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarkt-monitoring"
version = "0.1.0"
description = "Monitor new wallets placing high-value bets on EVM betting contracts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "web3>=6.15.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
polymarkt-monitor = "polymarkt_monitoring.main:cli_entrypoint"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
